backend/pyproject.toml
[project]
name = "Python-Script-backend"
version = "0.1.0"
description = "Python Script Runner Backend Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huey>=2.5.2",
    "redis~=5.0.1",
    "fastapi~=0.111.0",
    "uvicorn[standard]~=0.29.0",
    "sqlalchemy~=2.0.29",
    "pydantic~=2.7.0",
    "python-dotenv~=1.0.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["models", "schemas", "services"]